*.mk configuration files modified for compatibility with latest version of Mentor Sourcery CodeBench Lite toolchain (https://sourcery.mentor.com/GNUToolchain/release2641). All detected issues fixed.

+++ src/cmd/basic/basic.c
@@ -1534,7 +1534,7 @@ int get_value()
         value = get_num();
     else {
         ++cmdptr;
-        switch (c) {
+        switch ((int)c) {
         case '(':                       /* nesting */
             value = eval_sub();
             break;
